RavenDB-5295 Support for multiple `select new {} into` in RavenLinqOptimizer. Moving RavenLinqPrettifier to a separate class.
